combine api_response and fetch_wrapper files
