+++ src/gitlab/index.ts
@@ -1,4 +1,3 @@
 export * from "./api_responses";
 export * from "./merge_request_api";
 export * from "./uri";
-export * from "./fetch_wrapper";
